requirements.txt
requests==2.31.0
openai==0.28.0
python-dotenv==1.0.0
streamlit==1.26.0
streamlit-chat==0.1.1
pandas==2.1.0
